- Grid row/col resizing with double click on label - auto sizing of edit field
git-svn-id: svn://svn.pgadmin.org/trunk@2267 a7884b65-44f6-0310-8a51-81a127f17b15

+++ src/schema/pgObject.cpp
@@ -10,7 +10,6 @@
 
 // wxWindows headers
 #include <wx/wx.h>
-#include <wx/treectrl.h>
 
 // App headers
 #include "pgAdmin3.h"
@@ -665,7 +664,7 @@ gotToken:
             }
             if (tp->doBreak)
             {
-                fc += wxT("\n") + wxString(' ', indent);
+                fc += wxT("\n") + wxString(' ', (size_t)indent);
                 position = indent;
             }
             else
